Se puede exportar a csv delimitado por coma o por punto coma

+++ general/procesadores/GeneradorCSV.process.php
@@ -1,5 +1,5 @@
 <?php 
-ession_start();
+session_start();
 if (!isset($_SESSION['username'])){
   exit("<a href='../../index.php' ><img src='../images/401.png'>Iniciar Sesion </a>");
   
@@ -7,13 +7,14 @@ if (!isset($_SESSION['username'])){
 
 if(isset($_REQUEST["Opcion"])){
     $myPage="GeneradorCSV.php";
-    include_once("../../modelo/php_conexion.php");
+    include_once("../clases/administrador.class.php");
     
-    session_start();    
+       
     $idUser=$_SESSION['idUser'];
-    $obVenta = new conexion($idUser);
+    $obCon = new Administrador($idUser);
     
-    $OuputFile="../exports/";
+    $DatosRuta=$obCon->DevuelveValores("configuracion_general", "ID", 1);
+    $OuputFile=$DatosRuta["Valor"];
     $Link1=substr($OuputFile, -17);
     $Link="../../".$Link1;
     //print($Link);
@@ -23,46 +24,47 @@ if(isset($_REQUEST["Opcion"])){
     
     switch ($Opcion){
         case 1: //Exportar CSV 
+            if(file_exists($Link)){
+                unlink($Link);
+            }
             
-            unlink($Link);
-            $Tabla=$obVenta->normalizar(base64_decode($_REQUEST["TxtT"]));
-            $statement=$obVenta->normalizar(urldecode($_REQUEST["TxtL"]));
-            $Columnas=$obVenta->ShowColums($Tabla);
+            $Tabla=$obCon->normalizar($_REQUEST["Tabla"]);
+            
+            $Condicion=$obCon->normalizar($_REQUEST["Condicion"]);
+            $OrdenColumna=$obCon->normalizar($_REQUEST["OrdenColumna"]);
+            $AscDesc=$obCon->normalizar($_REQUEST["Orden"]);
+            $Separador=$obCon->normalizar($_REQUEST["Separador"]);
+            
+            $ColumnasSeleccionadas=$obCon->getColumnasVisibles($Tabla, "");  
+            $idTabla=$ColumnasSeleccionadas["Field"][0];        
+            if($Condicion<>""){
+                $Condicion=" WHERE ".$Condicion;
+            }
+            if($OrdenColumna==''){
+                $OrdenColumna=$idTabla;
+            }
+            
+            $Orden=" ORDER BY $OrdenColumna $AscDesc ";
+            
+            
+            
             $sqlColumnas="SELECT ";
             $CamposShow="";
-            foreach($Columnas["Field"] as $Campo){
-                $consulta=$obVenta->ConsultarTabla("tablas_campos_control", "WHERE NombreTabla='$Tabla' AND Campo='$Campo'");
-                $DatosCampo=$obVenta->FetchArray($consulta);
-                if($DatosCampo["Visible"]=='' or $DatosCampo["Visible"]=='1'){
-                    $sqlColumnas.="'$Campo',";
-                    $CamposShow.="`$Campo`,";
-                }
-                
+            foreach($ColumnasSeleccionadas["Field"] as $key => $value){
+                $Titulo= utf8_encode($ColumnasSeleccionadas["Visualiza"][$key]);                
+                $sqlColumnas.="'$Titulo' ,";
+                $CamposShow.=" CONVERT(`$value` USING utf8mb4),"; 
             }
             $sqlColumnas=substr($sqlColumnas, 0, -1);
             $CamposShow=substr($CamposShow, 0, -1);
             $sqlColumnas.=" UNION ALL ";
-            $Indice=$Columnas["Field"][0];
-            //print($Indice);
-            $sql=$sqlColumnas."SELECT $CamposShow FROM $statement INTO OUTFILE '$OuputFile' FIELDS TERMINATED BY ';' $Enclosed LINES TERMINATED BY '\r\n';";
-            $obVenta->Query($sql);
-            print("<a href='$Link' target='_top'><img src='../../images/download.gif'>Download</img></a>");
+            $Indice=$ColumnasSeleccionadas["Field"][0];
+            
+            $sql=$sqlColumnas."SELECT $CamposShow FROM $Tabla $Condicion INTO OUTFILE '$OuputFile' FIELDS TERMINATED BY '$Separador' $Enclosed LINES TERMINATED BY '\r\n';";
+            $obCon->Query($sql);
+            print("<div id='DivImagenDescargarTablaDB'><a href='$Link' download='$Tabla.csv' target='_top' style='text-align:center;position: absolute;top:50%;left:50%;padding:5px;' onclick=document.getElementById('DivImagenDescargarTablaDB').style.display='none';><h1>Descargar: </h1><img src='../../images/descargar3.png'></img></a></div>");
             break;
-        case 2:   //Resumen de facturacion agrupado por referencia en un periodo de tiempo
-            unlink($Link);
-            $FechaIni=$obVenta->normalizar($_REQUEST["TxtFechaIni"]);
-            $FechaFin=$obVenta->normalizar($_REQUEST["TxtFechaFin"]);
-            $sql="SELECT 'ID', 'FECHA', 'REFERENCIA','NOMBRE', 'DEPARTAMENTO', 'SUB1', 'SUB2','SUB3', 'SUB4','SUB5', 'CANTIDAD','TOTAL VENTA','COSTOS','EXISTENCIA' UNION ALL ";
-            $sql.="SELECT ID,`FechaFactura`, `Referencia`,`Nombre`,`Departamento`,`SubGrupo1`,`SubGrupo2`,"
-                . "`SubGrupo3`,`SubGrupo4`,`SubGrupo5`,SUM(`Cantidad`) as Cantidad,round(SUM(`TotalItem`)) as TotalVenta,"
-                . "round(SUM(`SubtotalCosto`)) as Costo,(SELECT Existencias FROM productosventa WHERE productosventa.Referencia=facturas_items.Referencia) as Existencias"
-                    . " FROM `facturas_items` "
-                . "WHERE `FechaFactura`>='$FechaIni' AND `FechaFactura`<='$FechaFin' "
-                . "GROUP BY `Referencia` "
-                . " INTO OUTFILE '$OuputFile' FIELDS TERMINATED BY ';' $Enclosed LINES TERMINATED BY '\r\n';";
-            $obVenta->Query($sql);
-            print("Resumen de Facturacion Generado exitosamente <a href='$Link'>Abrir</a>");
-            break;
+        
         }
 }else{
     print("No se recibió parametro de opcion");
